Refactored MessageProducer
