+++ .idea/workspace.xml
@@ -5,8 +5,14 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="31d0030a-af70-4981-9d40-cf9c9b5a2803" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/message_producer/kafka_producer.py" beforeDir="false" afterPath="$PROJECT_DIR$/message_producer/kafka_producer.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/message_producer/producer_interface.py" beforeDir="false" afterPath="$PROJECT_DIR$/message_producer/producer_interface.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/embedding_generator/DAO/ImageEmbeddingDao.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/embedding_generator/DAO/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/message_consumer/management/commands/consume_message.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/embedding_generator/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/embedding_generator/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/embedding_generator/task.py" beforeDir="false" afterPath="$PROJECT_DIR$/embedding_generator/task.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/message_consumer/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/message_consumer/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/message_consumer/callbacks.py" beforeDir="false" afterPath="$PROJECT_DIR$/message_consumer/callbacks.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/message_consumer/kafka_consumer.py" beforeDir="false" afterPath="$PROJECT_DIR$/message_consumer/kafka_consumer.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -227,27 +233,27 @@
         </line-breakpoint>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
           <url>file://$PROJECT_DIR$/embedding_generator/task.py</url>
-          <line>27</line>
+          <line>42</line>
           <option name="timeStamp" value="13" />
         </line-breakpoint>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
           <url>file://$PROJECT_DIR$/embedding_generator/task.py</url>
-          <line>34</line>
+          <line>49</line>
           <option name="timeStamp" value="15" />
         </line-breakpoint>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
           <url>file://$PROJECT_DIR$/embedding_generator/task.py</url>
-          <line>13</line>
+          <line>28</line>
           <option name="timeStamp" value="17" />
         </line-breakpoint>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
           <url>file://$PROJECT_DIR$/message_consumer/callbacks.py</url>
-          <line>111</line>
+          <line>113</line>
           <option name="timeStamp" value="27" />
         </line-breakpoint>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
           <url>file://$PROJECT_DIR$/message_producer/kafka_producer.py</url>
-          <line>39</line>
+          <line>32</line>
           <option name="timeStamp" value="28" />
         </line-breakpoint>
       </breakpoints>
